+++ noodles-bgzf/src/async/writer/deflate.rs
@@ -6,9 +6,10 @@ use std::{
 };
 
 use bytes::BytesMut;
-use flate2::Compression;
 use pin_project_lite::pin_project;
 use tokio::task::JoinHandle;
+
+use super::CompressionLevel;
 
 // (CDATA, CRC32, ISIZE)
 pub type GzData = (Vec<u8>, u32, u32);
@@ -21,9 +22,9 @@ pin_project! {
 }
 
 impl Deflate {
-    pub fn new(data: BytesMut, compression: Compression) -> Self {
+    pub fn new(data: BytesMut, compression_level: CompressionLevel) -> Self {
         Self {
-            handle: tokio::task::spawn_blocking(move || deflate(data, compression)),
+            handle: tokio::task::spawn_blocking(move || deflate(data, compression_level)),
         }
     }
 }
@@ -36,7 +37,7 @@ impl Future for Deflate {
     }
 }
 
-fn deflate(data: BytesMut, compression: Compression) -> io::Result<GzData> {
+fn deflate(data: BytesMut, compression_level: CompressionLevel) -> io::Result<GzData> {
     use crate::writer::deflate_data;
-    deflate_data(&data, compression)
+    deflate_data(&data, compression_level)
 }